added led

+++ inc/ultrasonic.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 /**
  * Represents an ultrasonic sensor
  * like HC-SR04
@@ -7,14 +8,14 @@
 class Ultrasonic {
     public:
     Ultrasonic(){}
-    Ultrasonic(int _pinTrigger, int _pinEcho);
+    Ultrasonic(uint8_t _pinTrigger, uint8_t _pinEcho);
     ~Ultrasonic(){}
 
     double distance(void);
 
     private:
-    int pinTrigger;
-    int pinEcho;
+    uint8_t pinTrigger;
+    uint8_t pinEcho;
 
     double timeToDistanceCM(double timeS);
     void trigger(void);
